Add solutions for day 24
Brute force + some memoization, though my solution for part 1 is smack
in the middle of the input, so I dunno how much the memoization is
helping. Solution for part 2 is near the beginning, so we hit it in
around 3 seconds in debug mode. In release mode, we can find both
solutions in under 11 seconds together, but it still feels slow.
There's probably somewhere I'm duplicating work, or generating inputs
that couldn't possibly work.